refactor(os): move cd from verticals.secret.repo to utils.os
