+++ tests/unit/cmd/scan/test_prereceive.py
@@ -7,8 +7,8 @@ from ggshield.__main__ import cli
 from ggshield.core.errors import ExitCode
 from ggshield.core.scan import StringScannable
 from ggshield.utils.git_shell import EMPTY_SHA, Filemode
+from ggshield.utils.os import cd
 from ggshield.verticals.secret import Result, Results, SecretScanCollection
-from ggshield.verticals.secret.repo import cd
 from tests.repository import Repository, create_pre_receive_repo
 from tests.unit.conftest import (
     _SIMPLE_SECRET_PATCH,
