+++ tests/unit/verticals/secret/test_scan.py
@@ -8,9 +8,8 @@ from pygitguardian.models import MultiScanResult, ScanResult
 from ggshield import __version__
 from ggshield.core.cache import Cache
 from ggshield.core.scan import Commit, ScanContext, ScanMode
-from ggshield.utils.os import get_os_info
+from ggshield.utils.os import cd, get_os_info
 from ggshield.verticals.secret import SecretScanner
-from ggshield.verticals.secret.repo import cd
 from tests.unit.conftest import UNCHECKED_SECRET_PATCH
